requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
Deprecated==1.2.13
Django==4.1.7
docker==6.0.1
exceptiongroup==1.1.1
fake-useragent==1.2.0
filelock==3.12.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.0
outcome==1.2.0
packaging==23.0
parse==1.19.1
parsel==1.8.1
prompt-toolkit==3.0.38
Protego==0.2.1
psycopg2==2.9.6
psycopg2-binary==2.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
pyOpenSSL==23.1.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
redis==4.1.4
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
retrying==1.3.4
scrapinghub==2.4.0
Scrapy==2.8.0
selenium==4.9.1
service-identity==21.1.0
shub==2.14.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.9
sqlparse==0.4.3
tldextract==3.4.1
toml==0.10.2
tqdm==4.55.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
tzlocal==2.1
urllib3==1.26.15
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
